+++ CHANGELOG.md
@@ -1,13 +1,26 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.1.1] -  2021-11-12
+### Changed
+- Switch E coli Malawi carriage tree to non ASC 
+- Switch E coli global tree to non ASC, using full Horesh collection
+- Use ClermonTyping v20.3 instead of isPcr for phylogroups
+- switch E coli ST410 and ST167 trees to non-ASC
+- add E coli ST131 tree and metadata
+### Fixed 
+- Updated kleb Malawi carriage tree to non ASC - 1.1.0 didn't have correct ones
+
 
 ## [1.1.0] - 2021-11-01
 ### Changed
 - Switch Klebsiella Malawi carriage tree to be without ASC
 - Switch Kleb global tree to be without ASC 
 - Switch Kleb Malawi all isolate tree to be without ASC
+- Switch to semantic versioning
 
 ## [1.0.0.3] - 2021-10-12
 ### Fixed